+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 ## Change log
 ----------------------
 
+Version 1.2
+-------------
+
+CHANGED:
+
+- changed to new package io.github.astrapi69
+- upgrade gradle version to 6.9
+- update of test dependency jobj-contract-verifier version to 3.4
+- update of test dependency silly-collections version to 8.6
+- update of test dependency test-objects version to 5.4
+- update of test dependency commons-lang3 version to 3.12.0
+- update of test dependency testng version to 7.4.0
+
 Version 1.1
 -------------